feature: add-static-filters-in-contact-group[WTEL-6934](https://webitel.atlassian.net/browse/WTEL-6934)
