+++ src/modules/configuration/modules/lookups/modules/contact-groups/components/the-contact-groups.vue
@@ -59,14 +59,6 @@
           @close="closeCreateGroupPopup"
         />
 
-        <add-contact-in-group-popup
-          :group-ids="staticGroupIds"
-          :namespace="namespace"
-          :shown="isShowAddContactInGroupPopup"
-          @close="isShowAddContactInGroupPopup = false"
-          @load-data="loadData"
-        />
-
         <div
           class="table-section__table-wrapper"
         >
@@ -174,7 +166,6 @@ import { useRouter } from 'vue-router';
 import { WebitelContactsGroupType } from 'webitel-sdk';
 
 import { useUserAccessControl } from '../../../../../../../app/composables/useUserAccessControl';
-import AddContactInGroupPopup from '../modules/contacts/components/add-contact-in-group-popup.vue';
 import CreateContactGroupPopup from './create-contact-group-popup.vue';
 
 const baseNamespace = 'configuration/lookups/contactGroups';
