+++ src/modules/configuration/modules/lookups/modules/contact-groups/modules/contactList/api/ContactListAPI.ts
@@ -1,0 +1,7 @@
+import { contacts } from '@webitel/ui-sdk/src/api/clients/сontacts/index';
+
+const ContactListAPI = {
+  getList: contacts.getList,
+};
+
+export default ContactListAPI;
