+++ src/modules/configuration/modules/lookups/modules/contact-groups/modules/contactList/namespace.ts
@@ -1,0 +1,3 @@
+import { ContactGroupsNamespace } from '../../namespace';
+
+export const ContactsListNamespace = `${ContactGroupsNamespace}/contacts-list`;
